changed route, animated edges, colored normal nodes
