+++ apps/ensadmin/src/app/@breadcrumbs/inspector/page.tsx
@@ -1,6 +1,6 @@
 import { BreadcrumbItem, BreadcrumbPage } from "@/components/ui/breadcrumb";
 
-export default function BreadcrumbsReactFlowPage() {
+export default function BreadcrumbsInspectorPage() {
   return (
     <BreadcrumbItem>
       <BreadcrumbPage>ENS Protocol Inspector (teaser)</BreadcrumbPage>
