+++ apps/ensadmin/src/app/inspector/custom-components/AnimatedSVGEdge.tsx
