+++ apps/ensadmin/src/app/inspector/custom-components/BaseNode.tsx
