+++ apps/ensadmin/src/app/inspector/custom-components/EdgeLabel.tsx
@@ -1,13 +1,14 @@
-export default function EdgeLabel({ transform, label }: { transform: string; label: string }) {
+export default function EdgeLabel({ transform, label, bgColor="#F7F9FB" }: { transform: string; label: string, bgColor?: string }) {
   return (
     <div
       style={{
         position: "absolute",
-        background: "#F7F9FB",
+        background: bgColor as string,
         padding: 0,
         color: "black",
         fontSize: 8,
         fontWeight: 500,
+          zIndex: 10,
         transform,
       }}
       className="nodrag nopan"
