+++ apps/ensadmin/src/app/inspector/custom-components/LabeledGroupNode.tsx
