+++ apps/ensadmin/src/app/inspector/page.tsx
@@ -1,0 +1,9 @@
+"use client";
+
+import dynamic from "next/dynamic";
+
+const InspectorClient = dynamic(() => import("./client"), { ssr: false });
+
+export default function InspectorPage() {
+  return <InspectorClient />;
+}
