+++ apps/ensadmin/src/app/inspector/schema-elements/edges.ts
@@ -15,6 +15,7 @@ export const initialEdges: Edge[] = [
       color: "black",
     },
     targetHandle: "StartInp",
+    animated: true
   },
   {
     id: "Namehash->UniRes",
@@ -29,6 +30,7 @@ export const initialEdges: Edge[] = [
       color: "black",
     },
     sourceHandle: "UniResOut",
+    animated: true
   },
   {
     id: "UniRes->ENSReg",
@@ -41,6 +43,7 @@ export const initialEdges: Edge[] = [
     style: {
       stroke: "black",
     },
+    animated: true
   },
   {
     id: "ENSReg->RegRec",
@@ -53,6 +56,7 @@ export const initialEdges: Edge[] = [
     style: {
       stroke: "black",
     },
+    animated: true
   },
   {
     id: "RegRec->ResolverL1",
@@ -66,6 +70,7 @@ export const initialEdges: Edge[] = [
       stroke: "black",
     },
     targetHandle: "RegRecInp",
+    animated: true
   },
   {
     id: "OffDatLok->Finish",
@@ -81,6 +86,7 @@ export const initialEdges: Edge[] = [
       color: "black",
     },
     sourceHandle: "FinishOut",
+    animated: true
   },
   {
     id: "ResolverL1->OffDatLok",
@@ -97,6 +103,7 @@ export const initialEdges: Edge[] = [
     },
     sourceHandle: "OffDatLokOut1",
     targetHandle: "ResolverL1Inp1",
+    animated: true
   },
   {
     id: "ResolverL1->OffDatLok(2)",
@@ -113,6 +120,7 @@ export const initialEdges: Edge[] = [
     },
     sourceHandle: "OffDatLokOut2",
     targetHandle: "ResolverL1Inp2",
+    animated: true
   },
   {
     id: "OffDatLok->ResolverL1",
@@ -129,6 +137,7 @@ export const initialEdges: Edge[] = [
     },
     targetHandle: "OffDatLokInp",
     sourceHandle: "ResolverL1Out",
+    animated: true
   },
   {
     id: "OffDatLok->CCIPRead",
@@ -145,6 +154,7 @@ export const initialEdges: Edge[] = [
     },
     targetHandle: "OffDatLokInp",
     sourceHandle: "CCIPReadOut",
+    animated: true
   },
   {
     id: "CCIPRead->OffDatLok",
@@ -153,6 +163,7 @@ export const initialEdges: Edge[] = [
     type: "start-end",
     data: {
       startLabel: "Return Data",
+      offsetY: -12,
     },
     markerEnd: {
       type: MarkerType.ArrowClosed,
@@ -160,6 +171,7 @@ export const initialEdges: Edge[] = [
     },
     sourceHandle: "OffDatLokOut",
     targetHandle: "CCIPReadInp",
+    animated: true
   },
   {
     id: "CCIPRead->ResolverL2",
@@ -176,6 +188,7 @@ export const initialEdges: Edge[] = [
     },
     sourceHandle: "ResolverL2Out",
     targetHandle: "CCIPReadInp",
+    animated: true
   },
   {
     id: "ResolverL2->CCIPRead",
@@ -193,5 +206,6 @@ export const initialEdges: Edge[] = [
     },
     targetHandle: "ResolverL2Inp",
     sourceHandle: "CCIPReadOut",
+    animated: true
   },
 ];
