--- apps/ensadmin/src/app/react-flow/page.tsx
@@ -1,9 +1,0 @@
-"use client";
-
-import dynamic from "next/dynamic";
-
-const ReactFlowClient = dynamic(() => import("./client"), { ssr: false });
-
-export default function ReactFlowPoCPage() {
-  return <ReactFlowClient />;
-}
